Change JSON deserializer for Resample2d and Resample1d to use u32 instead of usize; port render nodes to traits

+++ imageflow_core/src/flow/nodes/constrain.rs
@@ -147,8 +147,8 @@ impl NodeDefOneInputExpand for ConstrainDef{
 
             if resample {
                 let scale2d_params = s::Node::Resample2D {
-                    w: new_w as usize,
-                    h: new_h as usize,
+                    w: new_w,
+                    h: new_h,
                     up_filter: hints.and_then(|h| h.up_filter),
                     down_filter: hints.and_then(|h| h.down_filter),
                     scaling_colorspace: hints.and_then(|h| h.scaling_colorspace),
